feat(abi): release ABI 2.0 with unified wire and uncapped Scene budgets
Major ABI version 2; every service interface restarts at 1.0. Not
compatible with 1.x bundles.

Protocol
- Naming: verb-first methods, LOAD/UNLOAD, CREATE/DESTROY pairs; Timer
  methods drop the noun; PowerInfo service becomes Power.
- TEXTURE_LOAD absorbs TEXTURE_LOAD_ADAPTIVE; StreamingTexture is removed
  in favour of DynamicTexture.
- Field unification: byte_length -> length, src_* -> source_*, 32-bit
  font handles and input capabilities, service_info.handle ->
  service_handle, haptics_handle_t, reordered texture_info, scene node
  header and audio tone; gpio event payload drops the redundant device.
- Storage GET/REMOVE carry the key as trailing bytes like SET; devices
  LIST is paged (first_index/total_count).
- Raster uploads take explicit sizes on every entry and the SDK passes
  std::span.

Limits
- Policy macros leave the header; graphics_info reports max_text_bytes,
  max_scene_bytes, max_raster_bytes and max_surface_buffers, and the
  Guest runtime uses min(Host policy, static capacity).
- Scene nodes, containers and batch instances have no count cap: the
  wire's uint16 ids and node_count + batch_instance_count <= 65535 are
  the only bounds, storage grows per submit and OOM keeps the old scene.
- Label text (up to 1024 bytes) lives in a text arena on both sides:
  Host uses a two-half arena compacted before Apply, the compositor
  references it and is re-normalized after a compaction; Guest appends,
  compacts in place and grows on demand. Encoder scratch is reserved
  with the arrays it indexes so Submit never allocates.

Host, Guest runtime, SDK, conformance and Host unit tests updated; docs
describe the 2.0 behaviour. earth-garden stays a local benchmark.

+++ guest/sdk/input.hpp
@@ -16,11 +16,11 @@ class InputInfo final {
    private:
     static constexpr uint16_t kPressureCapability = 1U << 0U;
     static constexpr uint16_t kKeyEventsCapability = 1U << 1U;
-    constexpr InputInfo(uint16_t max_touch_points, uint16_t capabilities)
+    constexpr InputInfo(uint16_t max_touch_points, uint32_t capabilities)
         : max_touch_points_(max_touch_points), capabilities_(capabilities) {}
 
     uint16_t max_touch_points_{};
-    uint16_t capabilities_{};
+    uint32_t capabilities_{};
 
     friend class Input;
 };
